Experimental: disable baseband-xmm-power wakelock
This is probably a bad idea: that wakelock is surely there for a
reason, and removing it is bound to cause some kind of problems. But
after a little bit of testing, I haven't experienced the expected
problems, so I'm throwing this out there. It should improve idle
battery usage on tilapia if the other problems are bearable.